+++ asapdiscovery-docking/asapdiscovery/docking/scorer_v2.py
@@ -11,7 +11,7 @@ from asapdiscovery.data.postera.manifold_data_validation import TargetTags
 from asapdiscovery.data.schema_v2.ligand import LigandIdentifiers
 from asapdiscovery.data.schema_v2.target import TargetIdentifiers
 from asapdiscovery.docking.docking_v2 import DockingResult
-from asapdiscovery.ml.inference import InferenceBase, get_inference_cls_from_model_type
+from asapdiscovery.ml.inference.inference import InferenceBase, get_inference_cls_from_model_type
 from asapdiscovery.ml.models.ml_models import MLModelType
 from pydantic import BaseModel, Field
