--- asapdiscovery-ml/asapdiscovery/ml/inference/__init__.py
@@ -1,17 +1,0 @@
-from .inference import (
-    E3nnInference,
-    GATInference,
-    InferenceBase,
-    SchnetInference,
-    StructuralInference,
-    get_inference_cls_from_model_type,
-)
-
-__all__ = [
-    "InferenceBase",
-    "GATInference",
-    "StructuralInference",
-    "SchnetInference",
-    "E3nnInference",
-    "get_inference_cls_from_model_type",
-]
