--- asapdiscovery-ml/asapdiscovery/ml/models/__init__.py
@@ -1,17 +1,0 @@
-from .ml_models import (
-    ASAPMLModelRegistry,
-    LocalMLModelSpec,
-    MLModelBase,
-    MLModelRegistry,
-    MLModelSpec,
-    MLModelType,
-)
-
-__all__ = [
-    "ASAPMLModelRegistry",
-    "MLModelType",
-    "MLModelBase",
-    "MLModelSpec",
-    "LocalMLModelSpec",
-    "MLModelRegistry",
-]
