+++ asapdiscovery-ml/asapdiscovery/ml/scripts/train.py
@@ -29,12 +29,13 @@ from asapdiscovery.data.utils import (
     check_filelist_has_elements,
     extract_compounds_from_filenames,
 )
-from asapdiscovery.ml import (  # noqa: E402
+from asapdiscovery.ml.es import (  # noqa: E402
     BestEarlyStopping,
     ConvergedEarlyStopping,
-    GaussianNLLLoss,
-    MSELoss,
 )
+
+from asapdiscovery.ml.loss import GaussianNLLLoss, MSELoss
+
 from asapdiscovery.ml.utils import (
     build_dataset,
     build_model,
